pyproject.toml
[project]
name = "rag-series"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.0.12",
    "docx2txt>=0.9",
    "faster-whisper>=1.1.1",
    "groq>=0.26.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.7",
    "markdown>=3.10",
    "nltk>=3.9.2",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydub>=0.25.1",
    "pypdf>=5.6.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
    "tiktoken>=0.9.0",
    "transformers>=4.52.4",
    "unstructured[docx,pdf]>=0.18.21",
    "yt-dlp>=2025.5.22",
]
